backend/requirements.txt
# Django and Core Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Authentication and Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
django-oauth-toolkit==2.2.0

# API Documentation
drf-yasg==1.21.7
django-rest-swagger==2.2.0

# File Upload and Storage
Pillow==10.0.1
django-storages==1.14.2
boto3==1.34.0

# Caching and Sessions
redis==5.0.1
django-redis==5.4.0

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Payment Integration
stripe==7.6.0
paypalrestsdk==1.13.1
requests==2.31.0

# AI and ML Integration
openai==1.3.7
anthropic==0.7.7
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2

# Monitoring and Health Checks
django-health-check==3.17.0
prometheus-client==0.19.0

# Email and Notifications
django-anymail==10.2
twilio==8.10.0

# Image Processing
Pillow==10.0.1
python-magic==0.4.27

# Date and Time
python-dateutil==2.8.2
pytz==2023.3

# Environment and Configuration
python-decouple==3.8
django-environ==0.11.2

# Validation and Serialization
marshmallow==3.20.1
jsonschema==4.20.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-slugify==8.0.1
django-cleanup==8.1.0
django-safedelete==1.4.1

# Internationalization
django-modeltranslation==0.18.11

# Search
django-elasticsearch-dsl==8.0

# API Rate Limiting
django-ratelimit==4.1.0

# Logging
structlog==23.2.0

# Security
django-secure==1.0.2
django-axes==6.1.1

# Performance
django-cacheops==7.2
django-querycount==0.10.0 